+++ src/FacebookAds/Object/CustomAudience.php
@@ -34,7 +34,20 @@ use FacebookAds\Object\Values\CustomAudienceContentTypeValues;
 use FacebookAds\Object\Values\CustomAudienceCustomerFileSourceValues;
 use FacebookAds\Object\Values\CustomAudienceSubtypeValues;
 use FacebookAds\Object\Values\CustomAudienceTypes;
+use FacebookAds\Object\Fields\CustomAudienceMultikeySchemaFields;
 use FacebookAds\Object\CustomAudienceNormalizers\EmailNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\PhoneNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\MadidNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\GenderNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\BirthYearNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\DateNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\FirstNameNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\LastNameNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\FirstNameInitialNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\StateNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\CityNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\ZipNormalizer;
+use FacebookAds\Object\CustomAudienceNormalizers\CountryNormalizer;
 use FacebookAds\Object\CustomAudienceNormalizers\HashNormalizer;
 
 /**
@@ -571,6 +584,133 @@ class CustomAudience extends AbstractCrudObject {
 
       $payload['app_ids'] = $app_ids;
     }
+
+    return array('payload' => $payload);
+  }
+
+  /**
+  * @var \ArrayObject
+  */
+  protected $normalizers;
+
+  /**
+   * @return \ArrayObject
+   */
+  public function getNormalizers() {
+    if ($this->normalizers === null) {
+      $this->normalizers = new \ArrayObject(array(
+        new CustomAudienceNormalizers\EmailNormalizer(),
+        new CustomAudienceNormalizers\PhoneNormalizer(),
+        new CustomAudienceNormalizers\MadidNormalizer(),
+        new CustomAudienceNormalizers\GenderNormalizer(),
+        new CustomAudienceNormalizers\BirthYearNormalizer(),
+        new CustomAudienceNormalizers\DateNormalizer(),
+        new CustomAudienceNormalizers\FirstNameNormalizer(),
+        new CustomAudienceNormalizers\LastNameNormalizer(),
+        new CustomAudienceNormalizers\FirstNameInitialNormalizer(),
+        new CustomAudienceNormalizers\StateNormalizer(),
+        new CustomAudienceNormalizers\CityNormalizer(),
+        new CustomAudienceNormalizers\ZipNormalizer(),
+        new CustomAudienceNormalizers\CountryNormalizer(),
+      ));
+    }
+    return $this->normalizers;
+  }
+
+  /**
+   * Add users to the CustomAudiences with multiple keys. There is no max on the
+   * total number of users that can be added to an audience, but up to 10000
+   * users can be added at a given time.
+   *
+   * @param array $users
+   * @param array $types
+   * @param bool $is_hashed
+   * @param bool $is_normalized
+   * @return array
+   */
+  public function addUsersMultiKey(
+    array $users,
+    array $types,
+    $is_hashed = false,
+    $is_normalized = false) {
+    $params = $this->formatParamsMultiKey($users, $types, $is_hashed, $is_normalized);
+    return $this->getApi()->call(
+      '/'.$this->assureId().'/users',
+      RequestInterface::METHOD_POST,
+      $params)->getContent();
+  }
+
+
+  /**
+   * Delete users from AdCustomAudiences with multiple keys
+   *
+   * @param array $users
+   * @param array $types
+   * @param bool $is_hashed
+   * @param bool $is_normalized
+   * @return array
+   */
+  public function removeUsersMultiKey(
+    array $users,
+    array $types,
+    $is_hashed = false,
+    $is_normalized = false) {
+    $params = $this->formatParamsMultiKey($users, $types, $is_hashed, $is_normalized);
+    return $this->getApi()->call(
+      '/'.$this->assureId().'/users',
+      RequestInterface::METHOD_DELETE,
+      $params)->getContent();
+  }
+
+  /**
+   * Take users and format them correctly for the request
+   *
+   * @param array $users
+   * @param array $types
+   * @param bool $is_hashed
+   * @param bool $is_normalized
+   * @return array
+   */
+  protected function formatParamsMultiKey(
+    array $users,
+    array $types,
+    $is_hashed = false,
+    $is_normalized = false) {
+
+    if (!$is_hashed) {
+      if ($is_normalized) {
+        $normalizers = new \ArrayObject(array(
+          new HashNormalizer()
+        ));
+      }
+      else {
+        $normalizers = clone $this->getNormalizers();
+        $normalizers->append(new HashNormalizer());
+      }
+      foreach ($users as &$user) {
+        if (count($types) != count($user)) {
+          throw new \InvalidArgumentException(
+            "Number of keys in each list in the data should ".
+            "match the number of keys specified in scheme");
+          break;
+        }
+        foreach ($user as $index => &$key_value) {
+          $key = $types[$index];
+          foreach ($normalizers as $normalizer) {
+            if ($key_value &&
+                $key !== CustomAudienceMultikeySchemaFields::EXTERN_ID &&
+                $normalizer->shouldNormalize($key, $key_value)) {
+              $key_value = $normalizer->normalize($key, $key_value);
+            }
+          }
+        }
+      }
+    }
+
+    $payload = array(
+      'schema' => $types,
+      'data' => $users,
+    );
 
     return array('payload' => $payload);
   }