+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 
 ## Unreleased
+### Fixed
+ - Introduce `addUsersMultiKey` and `removeUsersMultiKey` in `CustomAudience` to still allow add users to `CustomAudience` with multiple keys after `CustomAudienceMultiKey` been deprecated.
+
+## v4.0.0
+### Changed
+- Graph API call upgrade to [v4.0](https://developers.facebook.com/docs/graph-api/changelog/version4.0)
 
 ## v3.3.1
 ### Changed
@@ -16,7 +22,7 @@ All notable changes to this project will be documented in this file.
 - Deprecated `parentID` in `AbstractCrudObject`.
 - Deprecated `CustomAudienceMultiKey`, use class `CustomAudience` instead.
 - Deprecated functions `create`, `read`, `update` in `AbstractCrudObject`. Check out our [recommended way](https://github.com/facebook/facebook-php-business-sdk#object-classes) to make API call.
-***`read` will reset the object fields, while `getSelf` will get a new object.*** For example : 
+***`read` will reset the object fields, while `getSelf` will get a new object.*** For example :
   ```
   $async_job = $adaccount->getInsightsAsync($fields, $params);
 
@@ -27,5 +33,3 @@ All notable changes to this project will be documented in this file.
     $async_job = $async_job->getSelf();
   }
   ```
-
-